[in_app_purchase] Rename the unified API (#1517)
Also moves the example keystore file into the right directory.
