+++ packages/in_app_purchase/example/lib/main.dart
@@ -5,7 +5,7 @@
 import 'dart:async';
 import 'dart:io';
 import 'package:flutter/material.dart';
-import 'package:in_app_purchase/in_app_purchase_connection.dart';
+import 'package:in_app_purchase/in_app_purchase.dart';
 
 void main() {
   runApp(MyApp());
