--- packages/in_app_purchase/lib/in_app_purchase_connection.dart
@@ -1,7 +1,0 @@
-// Copyright 2019 The Chromium Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-export 'src/in_app_purchase_connection/in_app_purchase_connection.dart';
-export 'src/in_app_purchase_connection/product_details.dart';
-export 'src/in_app_purchase_connection/purchase_details.dart';
