+++ packages/in_app_purchase/lib/src/in_app_purchase/product_details.dart
