+++ packages/in_app_purchase/lib/src/store_kit_wrappers/enum_converters.dart
@@ -1,6 +1,6 @@
 import 'package:in_app_purchase/store_kit_wrappers.dart';
 import 'package:json_annotation/json_annotation.dart';
-import 'package:in_app_purchase/in_app_purchase_connection.dart';
+import 'package:in_app_purchase/in_app_purchase.dart';
 
 part 'enum_converters.g.dart';
